change all values and returns to int

+++ bubbledocs-appserver/src/main/java/pt/tecnico/bubbledocs/content/Content.java
@@ -2,5 +2,5 @@ package pt.tecnico.bubbledocs.content;
 
 public abstract class Content {
 	
-	public abstract double getValue();
+	public abstract int getValue();
 }
