+++ bubbledocs-appserver/src/main/java/pt/tecnico/bubbledocs/content/Literal.java
@@ -2,13 +2,13 @@ package pt.tecnico.bubbledocs.content;
 
 public class Literal extends SimpleContent {
 
-	private double value;
+	private int value;
 	
-	public Literal(double value){
+	public Literal(int value){
 		this.value = value;
 	}
 	
-	public double getValue(){
+	public int getValue(){
 		return this.value;
 	}
 }
